requirements.txt
Flask==3.1.0
nltk==3.9.1
torch==2.4.1
transformers==4.39.3
tweet_preprocessor==0.6.0
